style: Extract custom Schedule button into his own file
